Implement pfc_memoryconfig + tests - WIP

+++ src/lib/include/pfc_memory.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-void * PFC_malloc(uint16_t size);
+void * PFC_malloc(uint32_t size);
 void PFC_free(void * ptr);
 
 char * PFC_strdup(const char * string);
