+++ src/lib/pfc_memory.c
@@ -5,7 +5,7 @@
 #include "pfc_memory.h"
 
 
-void * PFC_malloc(uint16_t size)
+void * PFC_malloc(uint32_t size)
 {
     void * ptr = malloc(size);
     memset(ptr, 0, size);
